Suggestion!
You have some code that you could reduce into for-loops. By turning number into an array, you'd be able to initialize all of the boxes with a for loop. This way, if you were to have an infinite number of boxes, it'd be covered in the for loop and you'd only have to change the max index instead of creating a new variable for each box. 

Lines 47-50 can be used to replace lines 52-88. 
Lines 232 to 244 can be used to replace lines 246-652

You might be able to use similar for-looping techniques in other areas of your code as well. Let me know if you want me to walk it over with you since this is a pretty common thing you'll run into as a programmer